worm/worm_rl_py/requirements.txt
cffi
psutil
python-nmap
paramiko
numpy